console: Refactor connect in organization events and device title section
